requirements.txt
fastapi
uvicorn
motor
odmantic
pydantic[dotenv]
bcrypt
python-jose[cryptography]
pymongo[srv]
pydantic
python-dotenv
PyJWT
pydantic[email]
passlib
pandas
openpyxl
python-multipart
geopy
httpx